Fix Channel model on client

+++ client/core/model/src/main/kotlin/io/github/vinicreis/pubsub/client/core/model/Channel.kt
@@ -7,7 +7,7 @@ data class Channel(
     val code: String,
     val name: String = code,
     val type: Type,
-    val pendingMessagesCount: Int,
+    val pendingMessagesCount: Long,
 ) {
     enum class Type {
         SIMPLE,
